+++ Core/CoreEvents.lua
@@ -21,4 +21,5 @@ addon.coreEvents = {
    ["ENCOUNTER_LOOT_RECEIVED"] = "OnEvent",
    ["BONUS_ROLL_RESULT"] = "OnBonusRoll",
    ["PLAYER_LOGOUT"] = "OnEvent",
+   ["ACTIVE_PLAYER_SPECIALIZATION_CHANGED"] = "OnSpecChanged",
 }