+++ Modules/Sync.lua
@@ -95,7 +95,7 @@ sync.syncHandlers = {
             addon:ActivateSkin(addon.db.profile.currentSkin)
         end,
         send = function()
-            return addon.db.profile
+            return addon:GetDBForExport()
         end
     },
     history = {