+++ __tests/wow_api.lua
@@ -446,6 +446,17 @@ function WoWAPI_FireEvent(event, ...)
 	end
 end
 
+--- ! NOT A WOW FUNCTION
+--- Emulating advancing time by firing WoWAPI_FireUpdate every 0.1 seconds.
+---@param seconds integer
+function _ADVANCE_TIME(seconds)
+	for i = 1, seconds do
+		for m = 0.1, 0.9, 0.1 do
+			WoWAPI_FireUpdate(_time + i + m)
+		end
+	end
+end
+
 function WoWAPI_FireUpdate(forceNow)
 	if forceNow then
 		_time = forceNow
@@ -482,6 +493,28 @@ C_Timer = {
 		end)
 	end,
 }
+
+local closureGeneration = {
+	function(f) return function(...) return f(...); end end,
+	function(f, a) return function(...) return f(a, ...); end end,
+	function(f, a, b) return function(...) return f(a, b, ...); end end,
+	function(f, a, b, c) return function(...) return f(a, b, c, ...); end end,
+	function(f, a, b, c, d) return function(...) return f(a, b, c, d, ...); end end,
+	function(f, a, b, c, d, e) return function(...) return f(a, b, c, d, e, ...); end end,
+};
+
+function GenerateClosure(f, ...)
+	local count = select("#", ...);
+	local generator = closureGeneration[count + 1];
+	if generator then
+		return generator(f, ...);
+	end
+	error("Closure generation does not support more than " .. (#closureGeneration - 1) .. " parameters");
+end
+
+function RunNextFrame(callback)
+	C_Timer.After(0, callback);
+end
 
 function GetServerTime()
 	return os.time()
@@ -596,6 +629,9 @@ strcmputf8i = function(a, b)
 	return a == b -- ~= nil
 end
 
+local function keyValueForTable(input)
+	return type(input) == "string" and "\"" .. input .. "\"" or tostring(input)
+end
 -- Not part of the WoWAPI, but added to emulate the ingame /dump cmd
 printtable = function(data, level)
 	if not data then return end
@@ -607,12 +643,15 @@ printtable = function(data, level)
 	for index, value in pairs(data) do
 		repeat
 			if type(value) ~= "table" then
-				print(ident .. "[" .. tostring(index) .. "] = " .. tostring(value) .. " (" .. type(value) .. ")");
+				print(ident ..
+				"[" ..
+				keyValueForTable(index) ..
+				"] = " .. keyValueForTable(value) .. ", -- (" .. type(value) .. ")");
 				break;
 			end
-			print(ident .. "[" .. tostring(index) .. "] = {")
+			print(ident .. "[" .. keyValueForTable(index) .. "] = {")
 			printtable(value, level + 1)
-			print(ident .. "}");
+			print(ident .. "},");
 		until true
 	end
 end