+++ ml_core.lua
@@ -61,6 +61,7 @@ function RCLootCouncilML:OnDisable()
 end
 
 function RCLootCouncilML:OnEnable()
+	self.Log "Enabled"
 	db = addon:Getdb()
 	self.lootTable = {} 		-- The MLs operating lootTable, see ML:AddItem()
 	self.oldLootTable = {}
@@ -139,7 +140,7 @@ function RCLootCouncilML:AddItem(item, bagged, slotIndex, owner, entry, boss)
 	entry.bagged = bagged
 	entry.lootSlot = slotIndex
 	entry.awarded = false
-	entry.owner = owner or addon.bossName
+	entry.owner = owner
 	entry.boss = boss or addon.bossName
 	entry.isSent = false
 	entry.typeCode = addon:GetTypeCodeForItem(item)
@@ -161,7 +162,7 @@ function RCLootCouncilML:AddItem(item, bagged, slotIndex, owner, entry, boss)
 			wipe(entry)
 			self.Log:D("Couldn't find item info for ", item)
 			addon:Print(format(L["ML_ADD_ITEM_MAX_ATTEMPTS"], tostring(item)))
-			addon:GetActiveModule("sessionframe"):Show(self.lootTable)
+			self:ShowSessionFrame(self.lootTable)
 			return
 		end
 		self:ScheduleTimer("Timer", 0.05, "AddItem", item, bagged, slotIndex, owner, entry, boss)
@@ -170,6 +171,14 @@ function RCLootCouncilML:AddItem(item, bagged, slotIndex, owner, entry, boss)
 		entry.attempts = nil
 		addon:SendMessage("RCMLAddItem", item, entry)
 	end
+end
+
+---Show the session frame with the given lootTable or `self.lootTable`.
+---@param lootTable table Defaults to `self.lootTable`.
+---@param disableAwardLater boolean Defaults to `false`. See [SessionFrame:Show()](lua://RCSessionFrame.Show)
+function RCLootCouncilML:ShowSessionFrame(lootTable, disableAwardLater)
+	addon:CallModule("sessionframe")
+	addon:GetActiveModule("sessionframe"):Show(lootTable or self.lootTable, disableAwardLater)
 end
 
 --- Removes everything that doesn't need to be sent in the lootTable
@@ -182,10 +191,10 @@ function RCLootCouncilML:GetLootTableForTransmit(overrideIsSent)
 			copy[k] = nil
 		else
 			v.bagged = nil
+			v.lootSlot = nil
 			v.awarded = nil
 			v.classes = nil
 			v.isSent = nil
-			v.lootSlot = nil
 			v.link = nil
 			v.ilvl = nil
 			v.texture = nil
@@ -273,8 +282,7 @@ end
 function RCLootCouncilML:AddUserItem(item, username)
 	if type(tonumber(item)) == "number" or string.find(item, "item:") then -- Ensure we can handle it
 		self:AddItem(item, false, nil, username) -- The item is neither bagged nor in the loot slot.
-		addon:CallModule("sessionframe")
-		addon:GetActiveModule("sessionframe"):Show(self.lootTable)
+		self:ShowSessionFrame()
 	else
 		addon:Print(format(L["ML_ADD_INVALID_ITEM"], tostring(item)))
 	end
@@ -290,8 +298,7 @@ function RCLootCouncilML:SessionFromBags()
 	if db.autoStart then
 		self:StartSession()
 	else
-		addon:CallModule("sessionframe")
-		addon:GetActiveModule("sessionframe"):Show(self.lootTable, true)  -- Disable award later checkbox in the sessionframe
+		self:ShowSessionFrame(self.lootTable, true)
 	end
 end
 
@@ -1289,8 +1296,7 @@ function RCLootCouncilML:Test(items)
 	if db.autoStart then
 		addon:Print(L["Autostart isn't supported when testing"])
 	end
-	addon:CallModule("sessionframe")
-	addon:GetActiveModule("sessionframe"):Show(self.lootTable)
+	self:ShowSessionFrame()
 end
 
 -- Returns true if we are ignoring the item